refactor(dap): split variable parser and types into SRP crates
