+++ crates/perl-dap-variable-parser/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "perl-dap-variable-parser"
+version = "0.10.0"
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Parser for Perl debugger variable output"
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/EffortlessMetrics/perl-lsp"
+homepage = "https://github.com/EffortlessMetrics/perl-lsp"
+documentation = "https://docs.rs/perl-dap-variable-parser"
+keywords = ["perl", "dap", "debug", "variables", "parser"]
+categories = ["development-tools::debugging", "parser-implementations"]
+
+[dependencies]
+perl-dap-variable-types = { workspace = true }
+once_cell = "1.21"
+regex = "1.12"
+thiserror = "2.0"
+
+[lints]
+workspace = true
