+++ crates/perl-dap-variable-parser/README.md
@@ -1,0 +1,3 @@
+# perl-dap-variable-parser
+
+Parser for Perl debugger variable output, producing `PerlValue` trees.
