+++ crates/perl-dap-variable-types/Cargo.toml
@@ -1,26 +1,20 @@
 [package]
-name = "perl-dap-variables"
+name = "perl-dap-variable-types"
 version = "0.10.0"
 edition.workspace = true
 rust-version.workspace = true
 authors.workspace = true
-description = "Variable rendering for Perl DAP"
+description = "Core Perl DAP variable value types"
 readme = "README.md"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/EffortlessMetrics/perl-lsp"
 homepage = "https://github.com/EffortlessMetrics/perl-lsp"
-documentation = "https://docs.rs/perl-dap-variables"
-keywords = ["perl", "dap", "debug", "variables", "rendering"]
+documentation = "https://docs.rs/perl-dap-variable-types"
+keywords = ["perl", "dap", "debug", "variables", "types"]
 categories = ["development-tools::debugging", "development-tools"]
 
 [dependencies]
-perl-dap-variable-parser = { workspace = true }
-perl-dap-variable-types = { workspace = true }
 serde = { version = "1.0", features = ["derive"] }
-
-[dev-dependencies]
-perl-tdd-support = { workspace = true }
-serde_json = { workspace = true }
 
 [lints]
 workspace = true
