+++ crates/perl-dap-variable-types/README.md
@@ -1,0 +1,3 @@
+# perl-dap-variable-types
+
+Core value types for Perl DAP variable parsing and rendering.
